Fix: unexpected keyword argument 'has_fp16_weights'
When iterating over the list of named modules, if a Linear8BitLt module came before a non-8 bit
module, the kwargs would be permuted to include the 'has_fp16_weights' and other parameters, and
the next time it had to instantiate a 'torch.nn.Linear' object, the constructor would error out
because **kwargs contained keyword arguments that were not appropriate for the base constructor.

This fixes that by copying kwargs when sending the extra data to Linear8BitLt.