Proposal to use more of pytest features
As discussed in our meeting, I wanted to propose to use more pytest
features for testing instead of relying on unittest.

To demonstrate the changes, I rewrote test_config.py (took only 30 min).
As you can see (side-by-side view is best), there are no substantital
changes, the core logic is exactly the same. Test coverage is also 100%
identical. But there are a few advantages:

- Using pytest.mark.parametrize instead of looping in the tests leads to
  less nesting
- Also, so far, if the first element in the loop fails, the next ones
  are not tested, but with parametrize, they are still tested because they
  are separate tests
- Allows to use simple assert a == b, instead of self.assertEqual(a, b)
  etc.
- Use pytest fixture tmp_path instead of TemporaryDirectory to save a
  few keystrokes and nesting

Some of the more useful features of pytest, such as fixtures, are not
even used here.

Regarding ease of use for outside contributors, I think it's a draw,
some things are easier, some require a bit of pytest knowledge to
understand. At least according to Google trends, pytest seems to be more
popular than unittest, though, so I would expect more contributors to
have existing knowledge of pytest.

My proposal would be to make use of more of the pytest features from
here on out. I'm *not* proposing to do a complete rewrite just for the
sake of rewriting. Instead, when the tests are being worked on (e.g. the
items from the roadmap), we can add more pytest features bit by bit.

Let me know what you think.