[WIP] Make LoRA work with custom models
This is a basis for discussion and not yet ready to be merged.

There are surprisingly few changes necessary to make LoRA work with
custom models. Those changes mostly revolve around attributes or methods
that are present by default on transformers models but not necessarily
on custom models, and many of those changes are only required for tests.
This PR identifies those lines of code and fixes them so that they work
with custom models _without_ creating any backwards incompatiblity for
models that are already supported.

A new suite of tests is introduced that checks that LoRA works for
various kinds of custom models.

In _test_inference_safetensors, I removed the layers_to_transform config
setting.

There are still some TODOs for this PR, or possible future PRs:

- Some more tests need to be added
- Some tests are failing and I'm not sure why, take a look at
  tests_failing.py
- It could be useful to have an example and/or update the docs.
- Layer indexing doesn't really work with custom models, except if they
  have a very specific naming pattern. I'll do a small refactor of layer
  indexing, perhaps this also enable it to work better with custom models.
- Test if custom models also work with AdaLora (and IA³).