Fix embedding LoRA weights initialization (#681)
When init_lora_weights=False, embedding LoRA weights were initialized as
all zeros, resulting in LoRA becoming an idenitity transform. This is
inconsistent with other module types, where init_lora_weights=False
results in random initialization and thus a non-identity operation.

As init_lora_weights=False is just for internal testing, users should
not be affected by this change. In fact, I updated the doc of this
parameter to - hopefully - better reflect this.

There is no direct test for this change. However, there are tests
in #676 that will fail without this fix, so it is tested indirectly.