MNT: Refactor matching of module names
As discussed internally.

This is a refactor of the logic used to match the name of a
module (layer) according to the config. At the moment, this is only used
for LoRA but could also be applicable to new methods added later (like
IA³).

The main idea here is to move the core matching logic into a utility
function. This not only facilities the re-use this function in the
future, it also makes it easier to test the core matching logic. To this
end, a test suite is provided.

To the best of my knowledge, the matching logic itself was not changed,
i.e. this change should be backwards compatible.

For user convenience, I also added a method inspect_matching_modules to
LoraModel. Users can call this to do a quick check which modules would
be fine-tuned and which wouldn't. LMK if you think this is useful.

During testing, I came across a couple of examples that, at least to me,
seemed unclear with regard to whether they should be a match or not.
Those are marked with a TODO in the test_utils.py. Please check these
examples closely to confirm if the existing behavior is correct or not.