Init IA³ weights randomly when so configured
Right now, no matter what the value of init_ia3_weights, these weights
are always initialized to be 1 (i.e. identity transforms). With this
fix, when init_ia3_weights=False, the weights are initialized randomly.
This is a setting mostly used for testing, so this fix has no user
impact.