Fix failing test for stable diffusion
The first issue was that the parametrization for the test did not
propagate the config kwargs to the individual configs of text encoder
and unet. This is now fixed.

The second issue is that there seems to be a tiny amount of randomness
in the output, e.g. sometimes 1 pixel of 50000 being different after
disabling adapters. The test is now more robust to those tiny
deviations. I have tested it locally 10 times without error, so I think
it is now robust enough and not too flaky.