FIX: Removes warnings about unknown pytest marker (#715)
This is a low prio PR but it solves an annoyance.

Right now, when running tests, the output is spammed by messages like:

> PytestUnknownMarkWarning: Unknown pytest.mark.multi_gpu_tests - is
this a typo? ...

This makes it more difficult to see the actually relevant information.
This PR fixes this by registering the two pytest markers we use, thus
removing the warnings.