Use specified device_map to load adapters weight
When loading adapters weight with from_pretrained, the original implementation will load the weights on the first visible cuda device. If `device_map` is passed to kwargs in from_pretrained, it can be used to ensure the weights are loaded to that device. This makes GPU memory management cleaner as the base model and the adapter weights can be easily assigned to the same device.