Fix an issue with false error raised
An error was raised when layers_pattern are not None and target_modules
was a string. The idea was that if target_modules is a string,
layers_pattern is not used, so this should not silently pass. However,
layers_pattern can be != None even if the user didn't change the
default, because a default layers_pattern is used in that case.

An error is still being raised if layers_to_transform is not falsy.