Warn when disabling adapters and bias != 'none'
As discussed internally.

For LoRA, given that bias='all' or bias='none', when doing inference
with a model in the disable_adapter context, the output will not be
identical to the output of the base model. This may be surprising to
users. Therefore, a warning is given. Furthermore, the docstring has
been extended to reflect this fact.