TST: Extend many tests to run w/ 8bit quantization
Description

Right now, our test suite for quantized models is very haphazard, as
there are only a handful of tests in test_common_gpu.py. With this PR,
we would basically extend all our common tests to run once without
quantization (i.e. the status quo) and once with quantization. Since the
latter requires a GPU, that would only apply to the nightly tests for
now.

The required changes to the tests were relatively minor, namely taking
care of moving the model to the right device in case of quantization,
and using autocast when calling forward (not sure if we can somehow get
rid of that).

Fixes

For the tests to pass, we need the fix from #794 (I added a quick fix
for the time being).

Also, I think we should raise an error inside of add_weighted_adapter
when the model is quantized (same as we do when unloading), as we don't
support merging quantized weights. If I'm wrong here and it should work,
we need to implement get_delta_weight on the quantized layers.

Failing tests

Still, even with those fixes, many tests are still failing. The kinds of
errors I get are along the lines of:

TypeError: unsupported operand type(s) for +=: 'dict' and 'Tensor'
RuntimeError: The size of tensor a (40) must match the size of tensor b (37) at non-singleton dimension 0
RuntimeError: shape '[2, 3, 4]' is invalid for input of size 192

This makes me think that these are real bugs when using quantization and
we should investigate and fix them. This also means that we should not
merge the PR until the bugs are fixed, or else our nightly run will
fail.

Status

I would like to get some feedback on this. Do we want to add these extra
tests or not? They are a bit slow, since we almost double the number of
tests we run (and this doesn't even include 4bit), but since only the
nightly tests are affected, I think it's okay.

If we don't want these tests, we should still investigate the failing
tests and fix potential bugs.