MNT: Refactor tuner forward methods for simplicity
This is to be consistent with changes in #794

Description

The forward methods of several tuner layers were partly unnecessarily
nested, which makes them harder to read and which can conceal bugs (as
in #794). Therefore, these methods have been refactored to (hopefully)
be as readable as possible. Moreover, the different methods are now
coded as similar as possible between the different implementations.

On top of those changes, I made the following adjustments:

- added some type hints to the methods of the layers
- removed a comment about code being copied which I think is not
  necessary
- for the lora embedding layer, we sometimes used F.embedding(...) and
  sometimes nn.Embedding.forward(self, ...) -- now, we consistently use
  only F.embedding(...)
- for IA³ Linear8bitLT, apply dtype conversion regardless of whether
  self.is_feedforward is True or False (as discussed internally)
- for the definition of lora Linear4bit, we (implicitly) checked if bnb
  is available and if bnb 4bit is available, but it is enough to check the
  latter, as it calls the former internally -- now we only check the
  latter, saving us 1 level of indentation
- a few times, ModuleDict.update(<dict>) is called when <dict> has only
  1 single item -- I simplified this code to just assign that item, which
  is more readable and consistent with other nearby code

Comments

In general, I ensured that the changes I made in the forward methods
result in the exact same outcome as previously. Please ensure that I
achieved this while reviewing this PR. This is probably easier with a
side-by-side diff view and not the default inline diff view of GitHub.

While working on this, I found that a lot of code is essentially
duplicated, but I think we can accept that if we want to avoid adding
too many layers of abstraction/indirection.

Moreover, I discovered multiple potential inconsistencies between the
implementations:

- sometimes, dtypes are converted, sometimes not
- dtypes are not always converted on the same intermediate result
- sometimes, result.clone() is called, sometimes not

I have added a "# FIXME" comment each time I spotted those differences.
Before merging, all those comments should be addressed. This either
means that we fix the inconsistencies, or, if they are intended, to add
a comment to explain why they are intended.