Fix issue with unused adapter parameters w/ DDP
Resolves #899

Description

When using PyTorch DistributedDataParallel, there is an error if DDP
wraps parameters that are not participating in the calculation of the
loss. This is annoying, as we often have those types of parameters,
e.g.:

- ModulesToSaveWrapper has a copy of the original weights
- When using multiple adapters, with only one being active

This PR aims at fixing this issue by exposing an attribute (actually:
property) called _ddp_params_and_buffers_to_ignore. DDP uses this
attribute to ignore the given parameters, making the error disappear.

Implementation

Please let me know if there is a better way to implement this. The
current solution has a few drawbacks:

Besides having to do this in the first place, it is also annoying that
we have to implement this for all tuner layers, as they can differ when
it comes to what parameters to ignore (we don't have a consistent naming
scheme).

Also, our unit tests cannot cover this directly, as DDP requires a
multi-gpu setup. Instead, the test is only for the existince, and
correct value, of the _ddp_params_and_buffers_to_ignore attribute.
Writing a proper DDP test would require more effort but should be feasible.

We also rely on a private attribute here, so there is a danger of this
breaking with future PyTorch releases.

DONE/TODO:

- [x] ModulesToSaveWrapper
- [x] LoRA
- [ ] IA³
- [ ] AdaLora