DOC Section on common issues encountered with PEFT
So far, this section contains two issues, not using latest package
versions and issues with loading PEFT models. This is based on what I
feel are issues that are commonly brought up.

If this is helpful, I think in the future, we should sections on common
issues with quantization, correcting wrong expectations about memory use
and runtime performance, etc., which I feel often come up.