Make base_model.peft_config single source of truth (#921)
Resolves #802, #923

For the problem description, please check the first issue.

I went with solution 2, i.e. making the base_model.peft_config the
"single source of truth" for the PEFT configuration. That way, we
minimize the risk of diverging configurations.

This does not apply for prompt learning, where we don't have a
peft_config on the base model (which is just the normal model, not a
PEFT class).

I added a setter for peft_config but from my testing, it isn't being
used. It's only there for completeness.