Add GH action to run unit tests with torch.compile (#943)
The GitHub action works by setting an environment variable
PEFT_DEBUG_WITH_TORCH_COMPILE=1. This causes the tests to run with
torch.compile if get_peft_model is being used. The action is triggered
manually and requires to indicate the branch to run it on.

With this action, we should be able to incrementally add support for
torch.compile in PEFT without disrupting the existing tests. Once we
fully support torch.compile, we can think about adding to the tests by
default and to remove the code from this PR.