ENH Support Conv2D layers for IA3
Adds support for Conv2D layers to the IA³ tuner. Tests are added to
check that they work.

Notes:

Unfortunately, when unmerging the Conv2d IA³ layers, there is quite a
bit of rounding error. I had to increase the tolerances for this
specific test case to make the tests pass. I'm not 100% sure why this
is, but I could imagine that for Conv2d, small errors accumulate because
of the convolution operation.

I also added tests for IA³ Linear layers for the custom models, which
also pass. However, there is an error when using Conv1D. The reason is
that merging fails because there is a shape mismatch when
fan_in_fan_out=True (which is set automatically for Conv1D). I'm not
sure how this should be fixed.

I also noticed that I don't understand what the feedforward_modules
parameter does in IA³. AFAICT, it always has to be set for all IA³
layers, i.e. self.is_feedforward always must be True. If it cannot be
False, we could as well remove feedforward_modules and automatically set
is_feedforward to True.