TST: Fix broken save_pretrained tests (#969)
Resolves #968

As the linked issue mentions, the test_save_pretrained_selected_adapters
test was broken because the 2nd adapter would load the weight of the
default adapter, instead of its own weights. This was a pretty quick
fix.

However, this made me wonder why the test was passing beforehand when it
is loading the wrong weights, so I dug deeper.

The first issue I encountered was that for IA³, we did not set
init_ia3_weights=False. For this reason, all weights were set to 1.0. Of
course, in that case, it doesn't matter what weights are loaded, they're
all the same. Therefore, I changed init_ia3_weights=False.

However, this still doesn't explain why this worked for LoRA, which,
even without init_lora_weights=True, should have some completely random
weights which should not be the same.

The reason for that is that we used get_peft_model_state_dict to get the
state_dict we used for comparison. This function only returns the
weights of one of the adapters (in this case default), so the weights
for the new adapter were never compared at all. Thus I changed this so
that all weights are now compared.

However, this now caused the tests for prompt encoders to fail. For some
reason, the state_dict from prompt encoding models is not the same after
loading. At this point, I stopped investigating and just wrote an
exception for prompt encoding to use get_peft_model_state_dict instead
of comparing the whole state_dict. Any insights would be appreciated.

Finally, for completeness, I added some checks for the existence of the
files of the new adapter.