Add base model metadata to model card
Resolves #938

This PR adds the base model metadata, if present, to the model card.

On top of this, the code for creating the model card has been refactored
to use the huggingface_hub classes instead of doing ad hoc parsing and
writing. A consequence of this is that if no model card exists, the
default template will now be used, with a lot of placeholder text.

LMK if this is not desired.