TST Cover Conv1D&Conv1d requires_grad for LoRA&IA3
Extend test coverage to also cover checking requires_grad for Conv1D and
Conv2d for LoRA and IA3, as well as Embedding for LoRA.

Note: I initially thought we had a bug, so I wrote these tests to fix
the bug in TDD fashion. It turns out that there is no bug, but it's
still good to have the tests, so here they are. I added a comment to the
lines where I thought we had the bug.