Update the release checklist
As discussed, we wanted to make small amendments to the release process,
so that we have a 0.N.0 commit on main. I also adjusted the wording here
and there.

There were some open questions, which I have marked with a FIXME. Please
let me know how to proceed with those.